refactorying the exposure selction logic into the LightValueSelector class

+++ LightMeter/src/com/kaipic/lightmeter/lib/LightSensorRepo.java
@@ -3,7 +3,7 @@ package com.kaipic.lightmeter.lib;
 import java.util.HashMap;
 import java.util.Map;
 
-import static java.lang.Double.valueOf;
+import static java.lang.Float.valueOf;
 
 public class LightSensorRepo {
   private LightSensorFactory lightSensorFactory;
@@ -22,7 +22,7 @@ public class LightSensorRepo {
       sensors.put(type, sensor);
     }
     if (type == LightSensorType.MANUAL) {
-      ((ManualLightSensor) sensor).setEVByEVAt100(new ExposureValue(valueOf(sensorInfo).floatValue()));
+      ((ManualLightSensor) sensor).setEVByEVAt100(new ExposureValue(valueOf(sensorInfo)));
     }
     return sensor;
   }
